Module webchat::chatserver

Reexports

use std::prelude::v1::*;

use std;

use std::borrow::Borrow;

use std::io;

use std::sync::{mpsc, Arc, Mutex};

use std::thread;

use rustc_serialize::json;

Modules

ws

Enums

ChatAction
Represents a single, atomic action taken by a chat member.

Constants

WS_ADDR

Traits

Receiver
A trait for receiving data frames and messages.

Sender
A trait for sending data frames and messages.

Functions

client_thread
Each client thread waits for input (or disconnects) from its respective clients and relays the appropriate messages via the relay MPSC channel.

listen
Create the relay MPSC (multi-producer/single-consumer) channel, spawn the relay thread, then listen for WebSocket clients and spawn their threads.

relay_thread
The relay thread handles all ChatActions received on its MPSC channel by sending them out to all of the currently connected clients.

start
Spawn a WebSocket listener thread.