Module webchat::chatserver::ws

Reexports

use std::prelude::v1::*;

Modules

receiver
The default implementation of a WebSocket Receiver.

sender
The default implementation of a WebSocket Sender.

Structs

Connection
Represents a connection to the server that has not been processed yet.

Message
Represents a WebSocket message.

Server
Represents a WebSocket server which can work with either normal (non-secure) connections, or secure WebSocket connections.

WebSocketProtocol
Represents a Sec-WebSocket-Protocol header

Enums

Type
Valid types of messages (in the default implementation)

WebSocketStream
A useful stream type for carrying WebSocket connections.