Struct webchat::chatserver::ws::Connection

pub struct Connection<R, W> where W: Write, R: Read(R, W);

Represents a connection to the server that has not been processed yet.

Methods

impl<R, W> Connection<R, W> where R: Read, W: Write

fn read_request(self) -> Result<Request<R, W>, Error>

Process this connection and read the request.

impl Connection<WebSocketStream, WebSocketStream>

fn shutdown(&mut self, how: Shutdown) -> Result<(), Error>

Shuts down the currennt connection in the specified way. All future IO calls to this connection will return immediately with an appropriate return value.